refactor: fetch the workflow in a separate right drawer component to improve type safety
